delete test.cpp and move to main.cpp
